added debugging stmt to relocator, fixed bug in encrypter
modified the call to UI().toUser() in FileRelocator to show a more
informative message for each file relocated and whethere it was
successful or not.
Fixed the bug in encrypter. The bug was that in the creation of a file
object to be put in the encrypted folder the folder path was added to
the file name without the appropriate "\\" between and thus a different
file was referenced.